script-to-email/requirements.txt
PyYAML==6.0
black==22.3.0
python-dateutil==2.8.2